emp_utils_cog now removed
Now integrated into AdamBot definition, and EmbedPages and util handling is now present in embed_pages cog.
